+++ lib/fl_utilities.dart
@@ -5,4 +5,4 @@ library fl_utilities;
 export 'src/extensions/build_context_x.dart';
 export 'src/extensions/text_x.dart';
 export 'src/extensions/value_changed_x.dart';
-export 'src/widgets/custom_list_view.dart';
+export 'src/widgets/flex_item_list_view/modules.dart';
