+++ lib/src/extensions/text_x.dart
@@ -11,11 +11,9 @@ extension TextX on Text {
   ///
   /// [opacity] should be a value between 0.0 and 1.0.
   ///
-  /// {@tool snippet}
   /// ```dart
   /// Text('Hello World').applyOpacity(opacity: 0.5);
   /// ```
-  /// {@end-tool}
   ///
   /// WARNING: This method is create a new [Text] instance, so the [Text.key]
   /// will not be used. Try apply it on the [key] argument instead.
@@ -89,11 +87,9 @@ extension TextStyleX on TextStyle {
   ///
   /// [opacity] should be a value between 0.0 and 1.0.
   ///
-  /// {@tool snippet}
   /// ```dart
   /// TextStyle(color: Colors.red).applyOpacity(0.5);
   /// ```
-  /// {@end-tool}
   ///
   /// See also:
   /// - [TextX.applyOpacity], extension to apply opacity on [Text] widget.
